==> scripts/4_census_combine_datasets.ipynb <==
# Census Datasets 2011, 2016 & 2021 : Combine & Pre-Process

The script imports the combined census data files of each year, checks and corrects consistency issues in columns (Income brackets and SA3 Level values) and combines the census datat to create a single dataset

import pandas as pd
import numpy as np
import os

## Import Census Datasets

#### Census 2011

path = r"/Users/patel/Documents/CF-Data Anaylst Course/portfolio_projects/mbs_analysis/datasets/"

df_census_2011 = pd.read_pickle(
    os.path.join(path, "clean_datasets/census_data/2011_cenus_combined.pkl")
)
df_census_2011.info()

#### Census 2016

df_census_2016 = pd.read_pickle(
    os.path.join(path, "clean_datasets/census_data/2016_cenus_combined.pkl")
)
df_census_2016.info()

### Census 2021

df_census_2021 = pd.read_pickle(
    os.path.join(path, "clean_datasets/census_data/2021_cenus_combined.pkl")
)
df_census_2021.info()

## Data Consistency Across All Datasets

### Income

Income brackets are not consistent across 2011, 2016 and 2021. Before the census datasets can be stacked, following steps will be executed for each dataset

1. Income brackets need to be standardized so the number of income brackets are same across all datasets
2. Any blank values will be populated to 0
2. Income brackets will be renamed

Below is the income mapping used. Average income is annual average of lowest and highest range within the bracket 

| Income Brackets               | Average Income | Average Income Column Name       |
|-------------------------------|----------------|----------------------------------|
| $1-$199 ($1-$10,399)          | 5200           | average_income_$5200             |
| $200-$299 ($10,400-$15,599)   | 12999.5        | average_income_$13000            |
| $300-$399 ($15,600-$20,799)   | 18199.5        | average_income_$18200            |
| $400-$599 ($20,800-$31,199)   | 25999.5        | average_income_$26000            |
| $600-$799 ($31,200-$41,599)   | 36399.5        | average_income_$36400            |
| $800-$999 ($41,600-$51,999)   | 46799.5        | average_income_$46800            |
| $1,000-$1,249 ($52,000-$64,999) | 58499.5      | average_income_$58500            |
| $1,250-$1,499 ($65,000-$77,999) | 71499.5      | average_income_$71500            |
| $1,500-$1,999 ($78,000-$103,999) | 90999.5     | average_income_$91000            |
| $2,000-$2,999 ($104,000-$155,999) | 129999.5   | average_income_$130000           |
| $3,000-$3,499 ($156,000-$181,999) | 168999.5   | average_income_$169000           |
| $3,500 or more ($182,000 or more) |             | average_income_$200000+          |

#### Census 2011

df_census_2011.info()

# rename the columns
new_column_names = {
    "$1-$10,399": "average_income_$5200",
    "$10,400-$15,599": "average_income_$13000",
    "$15,600-$20,799": "average_income_$18200",
    "$20,800-$31,199": "average_income_$26000",
    "$31,200-$41,599": "average_income_$36400",
    "$41,600-$51,999": "average_income_$46800",
    "$52,000-$64,999": "average_income_$58500",
    "$65,000-$77,999": "average_income_$71500",
    "$78,000-$103,999": "average_income_$91000",
    "$104,000+": "average_income_$130000",
    "Negative income": "negative_income",
    "Nil income": "no_income",
    "Not stated": "not_stated",
    "Not applicable": "not_applicable",
}
df_census_2011 = df_census_2011.rename(columns=new_column_names)

# add high income brackets which are present in 2021 census but not in census 2011.
# This is so the data is consistent for merging

# df_census_2011["average_income_$200000+"] = np.nan -- commenting out average_income_$200000+

""" df_census_2011["average_income_$200000+"] = df_census_2011[
    "average_income_$200000+"
].astype("Int64") """

# casting to reinforce intgeter value for consistency
df_census_2011["average_income_$169000+"] = np.nan
df_census_2011["average_income_$169000+"] = df_census_2011[
    "average_income_$169000+"
].astype("Int64")

# checking the shape and new columns
print(df_census_2011.shape)
df_census_2011.head()

df_census_2011.info()

#### Census 2016

df_census_2016.info()

# combining the brackets to create single bracket so its consistent with census 2011 dataset
df_census_2016["average_income_$26000"] = (
    df_census_2016["$400-$499 ($20,800-$25,999)"]
    + df_census_2016["$500-$649 ($26,000-$33,799)"]
)

df_census_2016["average_income_$91000"] = (
    df_census_2016["$1,500-$1,749 ($78,000-$90,999)"]
    + df_census_2016["$1,750-$1,999 ($91,000-$103,999)"]
)

# drop income columns that were combined and new column was derived above
df_census_2016.drop(
    [
        "$400-$499 ($20,800-$25,999)",
        "$500-$649 ($26,000-$33,799)",
        "$1,500-$1,749 ($78,000-$90,999)",
        "$1,750-$1,999 ($91,000-$103,999)",
    ],
    inplace=True,
    axis=1,
)

# rename the columns
new_column_names = {
    "$1-$149 ($1-$7,799)": "average_income_$5200",
    "$150-$299 ($7,800-$15,599)": "average_income_$13000",
    "$300-$399 ($15,600-$20,799)": "average_income_$18200",
    "$650-$799 ($33,800-$41,599)": "average_income_$36400",
    "$800-$999 ($41,600-$51,999)": "average_income_$46800",
    "$1,000-$1,249 ($52,000-$64,999)": "average_income_$58500",
    "$1,250-$1,499 ($65,000-$77,999)": "average_income_$71500",
    "$2,000-$2,999 ($104,000-$155,999)": "average_income_$130000",
    "$3,000 or more ($156,000 or more)": "average_income_$169000+",
    "Negative income": "negative_income",
    "Nil income": "no_income",
    "Not stated": "not_stated",
    "Not applicable": "not_applicable",
}
df_census_2016.rename(columns=new_column_names, inplace=True)

# commenting out average_income_$200000+
""" df_census_2016["average_income_$200000+"] = np.nan
df_census_2016["average_income_$200000+"] = df_census_2016[
    "average_income_$200000+"
].astype("Int64") """
df_census_2016.head()

#### Census 2021

There are more SA3 values in 2016 and 2021 compared to 2011. Compare SA3 values in each dataset to find the inconsistencies

df_census_2021.info()

# combining the brackets to create single bracket so its consistent with census 2011 dataset
df_census_2021["average_income_$26000"] = (
    df_census_2021["$400-$499 ($20,800-$25,999)"]
    + df_census_2021["$500-$649 ($26,000-$33,799)"]
)

df_census_2021["average_income_$91000"] = (
    df_census_2021["$1,500-$1,749 ($78,000-$90,999)"]
    + df_census_2021["$1,750-$1,999 ($91,000-$103,999)"]
)

df_census_2021["average_income_$169000+"] = (
    df_census_2021["$3,000-$3,499 ($156,000-$181,999)"]
    + df_census_2021["$3,500 or more ($182,000 or more)"]
)

# drop income columns that were combined and new column was derived above
df_census_2021.drop(
    [
        "$400-$499 ($20,800-$25,999)",
        "$500-$649 ($26,000-$33,799)",
        "$1,500-$1,749 ($78,000-$90,999)",
        "$1,750-$1,999 ($91,000-$103,999)",
        "$3,000-$3,499 ($156,000-$181,999)",
        "$3,500 or more ($182,000 or more)"
    ],
    inplace=True,
    axis=1,
)

# rename the columns
new_column_names = {
    "$1-$149 ($1-$7,799)": "average_income_$5200",
    "$150-$299 ($7,800-$15,599)": "average_income_$13000",
    "$300-$399 ($15,600-$20,799)": "average_income_$18200",
    "$650-$799 ($33,800-$41,599)": "average_income_$36400",
    "$800-$999 ($41,600-$51,999)": "average_income_$46800",
    "$1,000-$1,249 ($52,000-$64,999)": "average_income_$58500",
    "$1,250-$1,499 ($65,000-$77,999)": "average_income_$71500",
    "$2,000-$2,999 ($104,000-$155,999)": "average_income_$130000",
    "Negative income": "negative_income",
    "Nil income": "no_income",
    "Not stated": "not_stated",
    "Not applicable": "not_applicable",
}
df_census_2021.rename(columns=new_column_names, inplace=True)

df_census_2021.info()

### SA3 Level

Its important to ensure that Statistical Area Level 3 are from the same edition (release 2016), before stacking the census data and merging with MBS and patient experience. 

MBS 2013-19 data use SA3 areas / codes use edition released in 2016. Census 2011 uses  SA3 areas/codes from edition 2011. This means SA3 codes in census 2011 will need to be updated to match the corresponding 2016 SA3 codes. 

Statistical Area Level 3 2011 to Statistical Area Level 3 2016 mapping provided was provided ABS. This mapping file is imported to update the values in 2011 census.

# extracting sa3 values in to a set. The set will remove duplicates.
sa3_2011 = set(df_census_2011["SA3"])
sa3_2016 = set(df_census_2016["SA3"])
sa3_2021 = set(df_census_2021["SA3"])

# since 2016 and 2021 have the same number of unique SA3 values, checking to see if there are any differences
sa3_diff = sa3_2021 == sa3_2016

print(sa3_diff)

No differences in SA3 2021 and 2016 list. This makes sense since census data used 2016 edition of SA3

# finding SA3 codes that exist in 2011 but not in 2016 census.
sa3_diff_11 = sa3_2011 - sa3_2016
print(sa3_diff_11)

Found 11 SA3 codes are in 2011 census but not in 2016. External investigation found that SA3 areas were folded into other suburbs or separated into multiple suburbs due to its population. When separated, new SA3 code and name was assigned. 

Examples of 2011 to 2016 SA3 codes conversions.

| 2011_SA3 | Name               | 2016_SA3 |
|----------|--------------------|----------|
| 50804    | Kimberley          | 51001    |
| 50805    | Mid-West           | 51104    |
| 10101    | Goulburn - Mulwaree | 10105    |
| 31604    | Nambour            | 31607    |
| 50806    | East Pilbara       | 51002    |
| 80102    | Belconnen          | 80101    |

#### Importing 2011 to 2016 Mapping File

# import file containing SA3 2011 to 2016 maping
df_sa3_mapping = pd.read_csv(
    os.path.join(path, "original_datasets/state_sal3/sa3_2011_to_2016_mapping.csv"),
    index_col=None,
)
df_sa3_mapping.head(5)

df_sa3_mapping.info()

# converting codes to strings to ensure there is consistency between census and mapping file columns
df_sa3_mapping["SA3_CODE_2011"] = df_sa3_mapping["SA3_CODE_2011"].astype("str")
df_sa3_mapping["SA3_CODE_2011"] = df_sa3_mapping["SA3_CODE_2011"].str.strip()
df_sa3_mapping["SA3_CODE_2016"] = df_sa3_mapping["SA3_CODE_2016"].astype("str")
df_sa3_mapping["SA3_CODE_2016"] = df_sa3_mapping["SA3_CODE_2016"].str.strip()

#### Census 2011 : Find & Replace SA3 codes

# Convert the missing SA3 codes in a set to string
sa3_diff_11_str = set(map(str, sa3_diff_11))
sa3_diff_11_str

# extract the mapping for missing SA3 codes in 2011
df_sa3_mapping_missing = df_sa3_mapping[
    df_sa3_mapping["SA3_CODE_2011"].isin(sa3_diff_11_str)
]
df_sa3_mapping_missing

# SA3_Code_2011 has multiple mappings for SA3 codes as areas have been separated to multiple suburbs.
# For the project we are going to pick the first instance of the duplicate.
df_sa3_mapping_missing = df_sa3_mapping_missing.drop_duplicates(
    subset=["SA3_CODE_2011"], keep="first", ignore_index=False
)
df_sa3_mapping_missing

# set SA3_CODE_2011 as index and extract corresponding SA3_CODE_2016 value.
# convert to dictionary for replacement function
df_mapping_dict = df_sa3_mapping_missing.set_index("SA3_CODE_2011")[
    "SA3_CODE_2016"
].to_dict()
df_mapping_dict

# replacing SA3 2011 with corresponding SA3 2016 codes for consistency
df_census_2011["SA3"] = df_census_2011["SA3"].replace(df_mapping_dict)

df_census_2011["SA3"].value_counts(dropna=False)

# investigating why SA3 code has 2 rows.
df_census_2011[df_census_2011["SA3"] == "80101"]

Expected SA3 code to have 2 rows since Cotter - Namadgi (80102) was folded in to an existing suburb Belconnen(80101). To ensure there is unique SA3s in census data, the rows will be combined.

# Grouping any replacement action resulting in more than 1 SA3 rows in census 2011 data
df_census_2011_new = df_census_2011.groupby(by=["SA3"], as_index=False).sum()

# checking SA3 80101 has 1 record and values were added.
df_census_2011_new[df_census_2011_new["SA3"] == "80101"]

df_census_2011_new[df_census_2011_new["SA3"] == "80102"]

# checking the counts
df_census_2011_new["SA3"].value_counts()

# checking difference in length in SA3 census 2011 and 2016. Expected the length to be different
sa3_2011_v1 = set(df_census_2011["SA3"])
sa3_2011_v1 = set(map(str.strip, sa3_2011_v1))
print(len(sa3_2011_v1))
print(len(sa3_2016))

# checking that the replacement has worked and all values exist in 2011 also exist in 2016
# expected result is empty set
sa3_diff_vals = sa3_2011_v1 - sa3_2016
print(sa3_diff_vals)

### Year

Year field need to be added to each dataset to recognise the census the data is for

df_census_2011_new["Year"] = "2011"
df_census_2011_new["Year"] = df_census_2011_new["Year"].astype("str")
df_census_2011_new.head()

df_census_2016["Year"] = "2016"
df_census_2016["Year"] = df_census_2016["Year"].astype("str")
df_census_2016.head()

df_census_2021["Year"] = "2021"
df_census_2016["Year"] = df_census_2016["Year"].astype("str")
df_census_2021.head()

## Combine Census Datasets

df_census_2011_new.shape

df_census_2016.shape

df_census_2021.shape

All datasets have 26 columns and 1066 in the final combined census dataset. Expected no blanks

# checking for common columns. Expected 26 columns to be in common
df_census_2011_new.columns.intersection(df_census_2016.columns)

df_census_2021.columns.intersection(df_census_2016.columns)

# vertically stacking each of the dataframe to create a new dataset with years 2013-22
df_census_combined_2011_16_21 = pd.concat(
    [df_census_2011_new, df_census_2016, df_census_2021],
    axis=0,
    ignore_index=True,
)
df_census_combined_2011_16_21.shape

df_census_combined_2011_16_21

df_census_combined_2011_16_21.info()
df_census_combined_2011_16_21.to_clipboard()

df_census_combined_2011_16_21.to_pickle(
    os.path.join(path, "clean_datasets/census_data/2011_16_21_census_combined.pkl")
)
