==> scripts/6_patient_experience_gp_clean.ipynb <==
# Clean & Format Patient Experience GP Services Dataset

GP services is survey data collected on patient's experience with GP service during regular hours. The data is all percentages of respondents of survey answer

import pandas as pd
import os

# import gp service survey data file
path = r"/Users/patel/Documents/CF-Data Anaylst Course/portfolio_projects/mbs_analysis/datasets/"

df_px_gp_2018_23 = pd.read_csv(
    os.path.join(path, "clean_datasets/patient_experience/2023-2018_gp_services.csv")
)
df_px_gp_2018_23.info()

df_px_gp_2018_23.head(5)

## Data Consistency & Cleaning

### Year

df_px_gp_2018_23["year"].value_counts()

Update mapping of years to be as below:  This is so the years are consistent with mbs and census datasets

- 2022-23 : 2023
- 2021-22 : 2022
-  2020-21 : 2021
- 2019-20 : 2020
- 2018-19 : 2019


# replacing year field to standardize
year_replacement = {
    "2022-23": "2023",
    "2021-22": "2022",
    "2020-21": "2021",
    "2019-20": "2020",
    "2018-19": "2019",
}

df_px_gp_2018_23["year"] = df_px_gp_2018_23["year"].replace(year_replacement)
df_px_gp_2018_23["year"].value_counts()

df_px_gp_2018_23.head(10)

# convert to int64 for consistency with census and mbs
df_px_gp_2018_23["year"] = df_px_gp_2018_23["year"].astype("int")
df_px_gp_2018_23.dtypes

### Age Columns

#### Age 15-24

df_px_gp_2018_23["15–24"].value_counts()

# datapoints containing #. Only 1
df_px_gp_2018_23[df_px_gp_2018_23["15–24"].str.contains("#")]

This value contain # prior to the percentage. # is an indication  to use the number with caution as it contains high margin of error. For the purpose of the project, the # will be removed and data points will be converted from object to numeric

# removing the # from any values before converting to float
df_px_gp_2018_23["15–24"] = df_px_gp_2018_23["15–24"].str.replace("#", "", regex=False)

df_px_gp_2018_23["15–24"] = pd.to_numeric(df_px_gp_2018_23["15–24"], errors="raise")
df_px_gp_2018_23["15–24"].value_counts(dropna=False)

#### Age 25–34

# check for # values in numbers
df_px_gp_2018_23[df_px_gp_2018_23["25–34"].str.contains("#")]

# removing the # from any values before converting to float
df_px_gp_2018_23["25–34"] = df_px_gp_2018_23["25–34"].str.replace("#", "", regex=False)

# checking if str.replaced has worked. 0 result means all # values have been replaced
df_px_gp_2018_23[df_px_gp_2018_23["25–34"].str.contains("#")]

# conbert to numeric and raise any error
df_px_gp_2018_23["25–34"] = pd.to_numeric(df_px_gp_2018_23["25–34"], errors="raise")
df_px_gp_2018_23["25–34"].value_counts(dropna=False)

df_px_gp_2018_23.info()

df_px_gp_2018_23["35–44"].value_counts(dropna=False)

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["35–44"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["35–44"] = pd.to_numeric(
    df_px_gp_2018_23["35–44"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["35–44"].value_counts(dropna=False)

df_px_gp_2018_23["35–44"].dtype

#### Age 65–74

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["65–74"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["65–74"] = pd.to_numeric(
    df_px_gp_2018_23["65–74"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["65–74"].value_counts(dropna=False)

#### Age 75–84

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["75–84"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["75–84"] = pd.to_numeric(
    df_px_gp_2018_23["75–84"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["75–84"].value_counts(dropna=False)

df_px_gp_2018_23.info()

#### Age 85 and over

df_px_gp_2018_23.rename(columns={"85 and over": "85+"}, inplace=True)
df_px_gp_2018_23.head()

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["85+"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["85+"] = pd.to_numeric(
    df_px_gp_2018_23["85+"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["85+"].value_counts(dropna=False)

### Gender

#### Males

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["Males"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["Males"] = pd.to_numeric(
    df_px_gp_2018_23["Males"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["Males"].value_counts(dropna=False)

#### Females

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["Females"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["Females"] = pd.to_numeric(
    df_px_gp_2018_23["Females"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["Females"].value_counts(dropna=False)

df_px_gp_2018_23.rename(columns={"Males": "males", "Females": "females"}, inplace=True)

### Region Type

#### Major Cities of Australia

# checking column names
df_px_gp_2018_23.columns

# renaming column to be shorter and remove spaces
df_px_gp_2018_23.rename(
    columns={"major cities of australia": "major_cities"}, inplace=True
)

df_px_gp_2018_23.info()

#### Inner Regional Australia 

# renaming column to be shorter and remove spaces
df_px_gp_2018_23.rename(
    columns={"inner regional australia": "inner_regional"}, inplace=True
)

df_px_gp_2018_23.columns

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["inner_regional"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["inner_regional"] = pd.to_numeric(
    df_px_gp_2018_23["inner_regional"].str.replace("#", "", regex=False), errors="raise"
)
df_px_gp_2018_23["inner_regional"].value_counts(dropna=False)

#### Outer Regional Remote Australia

df_px_gp_2018_23.columns

# renaming column to be shorter and remove spaces
df_px_gp_2018_23.rename(
    columns={"outer regional remote australia": "outer_regional_remote"}, inplace=True
)

df_px_gp_2018_23.columns

# check for number of rows containg #
df_px_gp_2018_23[df_px_gp_2018_23["outer_regional_remote"].str.contains("#")]

# remove the hash and convert values to numeric (float)
df_px_gp_2018_23["outer_regional_remote"] = pd.to_numeric(
    df_px_gp_2018_23["outer_regional_remote"].str.replace("#", "", regex=False),
    errors="raise",
)
df_px_gp_2018_23["outer_regional_remote"].value_counts(dropna=False)

### Summary Statistics

df_px_gp_2018_23.dtypes

df_px_gp_2018_23.describe()

df_px_gp_2018_23.shape

### Export to Pickle File

df_px_gp_2018_23.to_pickle(
    os.path.join(
        path, "clean_datasets/patient_experience/2023-2018_gp_services_cleaned.pkl"
    )
)
