==> scripts/7_patient_experience_gp_ah_clean.ipynb <==
# Clean & Format Patient Experience After Hours GP Services Dataset

GP services after hours is survey data collected on patient's experience with after hours GP service. The data is all percentages of respondents of survey answer

import pandas as pd
import os

# import gp after hours service survey data.
path = r"/Users/patel/Documents/CF-Data Anaylst Course/portfolio_projects/mbs_analysis/datasets/"

df_px_gp_ah_2018_23 = pd.read_csv(
    os.path.join(
        path, "clean_datasets/patient_experience/2022-23_gp_after_hours_services.csv"
    )
)
df_px_gp_ah_2018_23.info()

## Data Consistency & Cleaning

### Year

df_px_gp_ah_2018_23["year"].value_counts()

Update mapping of years to be as below:  This is so the years are consistent with mbs and census datasets

- 2022-23 : 2023
- 2021-22 : 2022
-  2020-21 : 2021
- 2019-20 : 2020
- 2018-19 : 2019


# replacing year field to standardize
year_replacement = {
    "2022-23": "2023",
    "2021-22": "2022",
    "2020-21": "2021",
    "2019-20": "2020",
    "2018-19": "2019",
}

df_px_gp_ah_2018_23["year"] = df_px_gp_ah_2018_23["year"].replace(year_replacement)
df_px_gp_ah_2018_23["year"].value_counts()

df_px_gp_ah_2018_23.head(10)

# convert to int64 for consistency with census and mbs
df_px_gp_ah_2018_23["year"] = df_px_gp_ah_2018_23["year"].astype("int")
df_px_gp_ah_2018_23["year"].dtypes

### Gender

# rename gender columns
df_px_gp_ah_2018_23.rename(columns={"Male": "males", "Female": "females"}, inplace=True)
df_px_gp_ah_2018_23.columns

### Age Columns

Some age columns contain # indicating these values have high margin of error. For analysis purposes the # value is removed. 

#### Age 15–24 

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["15–24"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["15–24"] = pd.to_numeric(
    df_px_gp_ah_2018_23["15–24"].str.replace("#", "", regex=False), errors="raise"
)

# checking the updates and NAs
df_px_gp_ah_2018_23["15–24"].value_counts(dropna=False)

#### Age 25–34

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["25–34"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["25–34"] = pd.to_numeric(
    df_px_gp_ah_2018_23["25–34"].str.replace("#", "", regex=False), errors="raise"
)

# checking the updates and NAs
df_px_gp_ah_2018_23["25–34"].value_counts(dropna=False)

#### Age 55–64

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["55–64"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["55–64"] = pd.to_numeric(
    df_px_gp_ah_2018_23["55–64"].str.replace("#", "", regex=False), errors="raise"
)

# checking the updates and NAs
df_px_gp_ah_2018_23["55–64"].value_counts(dropna=False)

#### Age 65–74

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["65–74"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["65–74"] = pd.to_numeric(
    df_px_gp_ah_2018_23["65–74"].str.replace("#", "", regex=False), errors="raise"
)

# checking the updates and NAs
df_px_gp_ah_2018_23["65–74"].value_counts(dropna=False)

#### Age 75–84

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["75–84"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["75–84"] = pd.to_numeric(
    df_px_gp_ah_2018_23["75–84"].str.replace("#", "", regex=False), errors="raise"
)

# checking the updates and NAs
df_px_gp_ah_2018_23["75–84"].value_counts(dropna=False)

#### Age 85 and over

# renaming the column
df_px_gp_ah_2018_23.rename(columns={"85 and over": "85+"}, inplace=True)

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["85+"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["85+"] = pd.to_numeric(
    df_px_gp_ah_2018_23["85+"].str.replace("#", "", regex=False), errors="raise"
)

# checking the updates and NAs
df_px_gp_ah_2018_23["85+"].value_counts(dropna=False)

### Region Type

#### Major Cities in Australia

# renaming the column
df_px_gp_ah_2018_23.rename(
    columns={"major cities of australia": "major_cities"}, inplace=True
)

#### Inner Regional in Australia

# renaming the column
df_px_gp_ah_2018_23.rename(
    columns={"inner regional australia": "inner_regional"}, inplace=True
)

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["inner_regional"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["inner_regional"] = pd.to_numeric(
    df_px_gp_ah_2018_23["inner_regional"].str.replace("#", "", regex=False),
    errors="raise",
)

# checking the updates and NAs
df_px_gp_ah_2018_23["inner_regional"].value_counts(dropna=False)

#### Outer Regional Remote Australia

# renaming the column
df_px_gp_ah_2018_23.rename(
    columns={"outer regional remote australia": "outer_regional_remote"}, inplace=True
)

# Check for # value in the column
df_px_gp_ah_2018_23[df_px_gp_ah_2018_23["outer_regional_remote"].str.contains("#")]

# remove # and convert to numeric value. Raise any errors for any failed ones
df_px_gp_ah_2018_23["outer_regional_remote"] = pd.to_numeric(
    df_px_gp_ah_2018_23["outer_regional_remote"].str.replace("#", "", regex=False),
    errors="raise",
)

# checking the updates and NAs
df_px_gp_ah_2018_23["outer_regional_remote"].value_counts(dropna=False)

df_px_gp_ah_2018_23.shape

df_px_gp_ah_2018_23.describe().to_clipboard()

### Export to Pickle File

# exporting to pickle file for further analysis
df_px_gp_ah_2018_23.to_pickle(
    os.path.join(
        path, "clean_datasets/patient_experience/2023-2018_gp_after_hours_cleaned.pkl"
    )
)
