==> scripts/10_census_mbs_combine_no_pivot.ipynb <==
# Combine: Cenus & MBS Datasets

This script combines non pivot version of MBS and Census datasets, making it easier to manage columns and EDA. The MBS dataset contains service level 1,2 and 3. Out of pocket by income % is calculated again here.

1. Imported Census data. Pivoted the demograghic columns into rows. Standardized the age values to be the same as MBS
2. Joined Cenus to MBS dataset
3. Derived OOP % by income for each income bracket.

import pandas as pd
import numpy as np
import os
import seaborn as sns

## Import Census and MBS Datasets

### Import Census

# import the census and assign to a dataframe

# setup path to original dataset
path = r"/Users/patel/Documents/CF-Data Anaylst Course/portfolio_projects/mbs_analysis/datasets/"
df_census_2014_22 = pd.read_pickle(
    os.path.join(
        path, "clean_datasets/census_data/2014_22_census_complelete_standardized.pkl"
    )
)
df_census_2014_22.info(10)

### Pivot Demograhic Columns to Rows

# defining columns to pivot
age_gender_p_columns = [
    "age_0-24",
    "age_25-44",
    "age_45-64",
    "age_65+",
    "male_pop",
    "female_pop",
    "total_population",
]
# defining columns to remain as is (id columns). expected these to duplicate
id_columns = [
    "key",
    "SA3",
    "Year",
    "negative_income",
    "no_income_or_not_applicable",
    "average_income_$5200",
    "average_income_$13000",
    "average_income_$18200",
    "average_income_$26000",
    "average_income_$36400",
    "average_income_$46800",
    "average_income_$58500",
    "average_income_$71500",
    "average_income_$91000",
    "average_income_$130000",
    "average_income_$169000+",
]

# using melt function to pivot the demograhic into rows, naming columns Population_demograhic. Values stored in Population
df_census_2014_22_pivot = df_census_2014_22.melt(
    id_vars=id_columns,
    value_vars=age_gender_p_columns,
    var_name="Population_demographic",
    value_name="Population",
)
df_census_2014_22_pivot

df_census_2014_22_pivot.info()

# checking values before and after pivots
df_census_2014_22[df_census_2014_22["key"] == "2014-10102"]

df_census_2014_22_pivot[df_census_2014_22_pivot["key"] == "2014-10102"]

# extracting unique values, which need to be updated to reflect same as MBS for join
df_census_2014_22_pivot["Population_demographic"].unique()

# Update demographic value to match the MBS dataset for left join
demographic_mapping = {
    "age_0-24": "0-24",
    "age_25-44": "25-44",
    "age_45-64": "45-64",
    "age_65+": "65+",
    "male_pop": "Males",
    "female_pop": "Females",
    "total_population": "All persons",
}
df_census_2014_22_pivot["Population_demographic"] = df_census_2014_22_pivot[
    "Population_demographic"
].map(demographic_mapping)
df_census_2014_22_pivot["Population_demographic"].unique()

df_census_2014_22_pivot[df_census_2014_22_pivot["key"] == "2014-10102"]

# import the mbs file and assign to a dataframe

df_mbs_2014_23 = pd.read_pickle(
    os.path.join(
        path, "clean_datasets/mbs_data/2014-22_mbs_state_complete_no_pivot.pkl"
    )
)
df_mbs_2014_23.info()

## Combine Census & MBS

Census will be combined with MBS using left join. Key columns will be Year, GeographicCode (SA3), DemograhicGroup (Population Demographic)

df_census_2014_22_pivot.shape

df_mbs_2014_23.shape

Expected 229,569 rows and 42 columns - number of rows will be the same as MBS. No nulls expected

# left join Cenus to MBS
df_mbs_census_combined_np = df_mbs_2014_23.merge(
    df_census_2014_22_pivot,
    how="left",
    left_on=["Year", "GeographicCode", "DemographicGroup"],
    right_on=["Year", "SA3", "Population_demographic"],
    indicator=True,
)
df_mbs_census_combined_np.shape

# checking for any left only joins
df_mbs_census_combined_np["_merge"].value_counts(dropna=False)

# checking for nulls
df_mbs_census_combined_np.isna().sum()

# checking specific key to ensure the join has worked
df_mbs_census_combined_np[
    (df_mbs_census_combined_np["key_x"] == "2014-10102")
    & (df_mbs_census_combined_np["Service"] == "Allied Health attendances (total)")
]

### Drop Unrequired Columns

# checking column that have duplicated or not required to drop
df_mbs_census_combined_np.columns

# dropping merge, key_y and SA3 columns
df_mbs_census_combined_np.drop(columns=["_merge", "key_y", "SA3"], inplace=True)
df_mbs_census_combined_np.columns

## Derive %OOP from Income

df_mbs_census_combined_np_new = df_mbs_census_combined_np.copy()

df_mbs_census_combined_np_new["%_out_of_pocket_by_$5200"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 5200
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$13000"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 13000
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$18200"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 18200
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$26000"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 26000
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$36400"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 36400
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$46800"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 46800
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$58500"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 58500
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$71500"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 71500
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$91000"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 91000
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$130000"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 130000
) * 100
df_mbs_census_combined_np_new["%_out_of_pocket_by_$169000+"] = (
    df_mbs_census_combined_np_new["Out_of_pocket_cost_per_person"] / 169000
) * 100

df_mbs_census_combined_np_new.head(3)

df_mbs_census_combined_np_new.isna().sum()

for col in df_mbs_census_combined_np_new.columns:
    if col.startswith("%_out_of_pocket_by"):
        df_mbs_census_combined_np_new[col] = df_mbs_census_combined_np_new[col].round(3)

df_mbs_census_combined_np_new.head(3)

df_mbs_census_combined_np_new.shape

### Export MBS Census Combined Dataset

df_mbs_census_combined_np_new.to_pickle(
    os.path.join(path, "clean_datasets/2014_22_mbs_cenus_combined_no_pivot.pkl")
)

#### Exporting SA3 & Names

sa_codes_names = df_mbs_census_combined_np_new[
    ["GeographicCode", "GeographicAreaName"]
].copy()
sa_codes_names.drop_duplicates(inplace=True)
sa_codes_names
sa_codes_names.to_pickle(
    os.path.join(path, "clean_datasets/geojson/sa3_name_mbs_list.pkl")
)
