==> scripts/16_mbs_service_oop_clean.ipynb <==
# MBS Out of Pocket Cost Per Service By State & Year Dataset

Alternate dataset providing avg out of pocket cost per service per state. 

The script cleans and transforms dataset prior to combining with master dataset. Exported to pickle and csv file. 

import pandas as pd
import os

## Import Dataset

path = r"/Users/patel/Documents/CF-Data Anaylst Course/portfolio_projects/mbs_analysis/datasets/"

df_oop_service = pd.read_csv(
    os.path.join(path, "original_datasets/mbs_data/state_year_oop_cost.csv")
)
df_oop_service.info()

## Cleaning Dataset

### Renaming Columns

# Renaming columns
df_oop_service.rename(
    columns={
        "Financial Year": "Year",
        "Broad Type of Service": "Service",
        " Avg Patient Contribution Per Service: Out of Hospital Patient Billed ": "Patient_OOP_per_service",
    },
    inplace=True,
)
df_oop_service

### Subseting to Extract 2014-2022 Dataset

df_oop_service["Year"].unique()

df_oop_service["Service"].unique()

df_oop_service_subset = df_oop_service[
    (
        df_oop_service["Year"].isin(
            [
                "2021-22",
                "2020-21",
                "2019-20",
                "2018-19",
                "2017-18",
                "2016-17",
                "2015-16",
                "2014-15",
                "2013-14",
            ]
        )
    )
    & df_oop_service["Service"].isin(
        [
            "Total GP Non-Referred Attendances",
            "Practice Nurse",
            "Other Allied Health",
            "Specialist Attendances",
            "Diagnostic Imaging",
        ]
    )
    & df_oop_service["State"].isin(
        ["NSW", "Vic", "Qld", "SA", "WA", "Tas", "NT", "ACT"]
    )
].copy()
df_oop_service_subset

df_oop_service_subset["Service"].unique()

df_oop_service_subset["State"].unique()

df_oop_service_subset["Year"].unique()

### Standardisation of Values

#### Year

year_replacement = {
    "2021-22": "2022",
    "2020-21": "2021",
    "2019-20": "2020",
    "2018-19": "2019",
    "2017-18": "2018",
    "2016-17": "2017",
    "2015-16": "2016",
    "2014-15": "2015",
    "2013-14": "2014",
}

df_oop_service_subset["Year"] = df_oop_service_subset["Year"].replace(year_replacement)

df_oop_service_subset["Year"].unique()

#### Service Names

service_name_replacement = {
    "Total GP Non-Referred Attendances": "GP attendances (total)",
    "Practice Nurse": "Nursing and Aboriginal Health Workers (total)",
    "Other Allied Health": "Allied Health attendances (total)",
    "Specialist Attendances": "Specialist attendances (total)",
    "Diagnostic Imaging": "Diagnostic Imaging (total)",
}
df_oop_service_subset["Service"] = df_oop_service_subset["Service"].replace(
    service_name_replacement
)

df_oop_service_subset["Service"].unique()

#### Updating Data Types

df_oop_service_subset["Year"] = df_oop_service_subset["Year"].astype("int")

df_oop_service_subset["Patient_OOP_per_service"] = df_oop_service_subset[
    "Patient_OOP_per_service"
].astype("float")

df_oop_service_subset.info()

## Export to Pickle & CSV Files

# exporting MBS-Census-State-Year-Service Level 1 (No demographic Data)
df_oop_service_subset.to_csv(
    os.path.join(
        path, "clean_datasets/cleaned_csv/mbs_oop_service_state_year_clean.csv"
    )
)

df_oop_service_subset.to_pickle(
    os.path.join(path, "clean_datasets/mbs_data/mbs_oop_service_state_year_clean.pkl")
)
