==> scripts/18_exporting_data_slices.ipynb <==
## Exporting Data Slices

Script was use to export various dataslieces for additional analysis in Excel and Tableau

import pandas as pd
import os
import numpy as np

pd.set_option("display.max_rows", 40)
pd.set_option("display.max_columns", None)

### Import Datasets

path = r"/Users/patel/Documents/CF-Data Anaylst Course/portfolio_projects/mbs_analysis/datasets/"

df_mbs_sa3_service_1_np = pd.read_pickle(
    os.path.join(
        path, "clean_datasets/2014-22_combined_service_1_categories_no_pivot.pkl"
    )
)
df_mbs_sa3_service_1_np.info()

df_mbs_sa3_service_1_np["Population_demographic"].value_counts()

df_mbs_sa3_service_1_np.columns

### Import MBS OOP Service Dataset

df_mbs_oop_service = pd.read_pickle(
    os.path.join(path, "clean_datasets/mbs_data/mbs_oop_service_state_year_clean.pkl")
)
df_mbs_oop_service.info()

### Combine MBS Census With MBS OOP.

df_mbs_oop_service["State"].unique()

df_mbs_sa3_service_1_np["StateTerritory"].unique()

df_mbs_sa3_service_1_np["Service"].unique()

df_mbs_oop_service["Service"].unique()

df_mbs_sa3_service_1_np_new = df_mbs_sa3_service_1_np.merge(
    df_mbs_oop_service,
    how="left",
    left_on=["Year", "StateTerritory", "Service"],
    right_on=["Year", "State", "Service"],
)

df_mbs_sa3_service_1_np_new.isnull().sum()

df_mbs_sa3_service_1_np_new.columns

# calculate oop per person annually.

df_mbs_sa3_service_1_np_new["OOP_per_person_annual"] = (
    df_mbs_sa3_service_1_np_new["No_of_service_per_person"]
    * df_mbs_sa3_service_1_np_new["Patient_OOP_per_service"]
)

df_mbs_sa3_service_1_np_exp = df_mbs_sa3_service_1_np_new[
    [
        "Year",
        "StateTerritory",
        "GeographicCode",
        "GeographicAreaName",
        "GeographicGroup",
        "Service",
        "DemographicGroup",
        "MBS_per_100",
        "No_of_patients",
        "No_of_services",
        "%_People_had_service",
        "Services_100_people",
        "Total_mbs_paid_$",
        "Total_provider_fees_$",
        "ERP",
        "Out_of_Pocket",
        "Out_of_pocket_cost_%",
        "Patient_OOP_per_service",
        "No_of_service_per_person",
        "OOP_per_person_annual",
        "negative_income",
        "no_income_or_not_applicable",
        "average_income_$5200",
        "average_income_$13000",
        "average_income_$18200",
        "average_income_$26000",
        "average_income_$36400",
        "average_income_$46800",
        "average_income_$58500",
        "average_income_$71500",
        "average_income_$91000",
        "average_income_$130000",
        "average_income_$169000+",
        "Population",
        "Service_Usage",
        "Service_Coverage",
    ]
]

# exporting MBS-Census-State-Year-Service Level 1 (No demographic Data)
df_mbs_sa3_service_1_np_exp.to_csv(
    os.path.join(
        path, "clean_datasets/cleaned_csv/mbs_sa3_service_1_categories_oop.csv"
    )
)

## State-Year MBS Data by Gender

df_mbs_sa3_gender = df_mbs_sa3_service_1_np_new[
    df_mbs_sa3_service_1_np_new["DemographicGroup"].isin(["Males", "Females"])
]
df_mbs_sa3_gender.shape

df_mbs_sa3_gender_state = df_mbs_sa3_gender.groupby(
    by=["Year", "StateTerritory", "Service", "DemographicGroup"], as_index=False
).agg(
    {
        "No_of_patients": "sum",
        "No_of_services": "sum",
        "Total_mbs_paid_$": "sum",
        "Total_provider_fees_$": "sum",
        "ERP": "sum",
        "Patient_OOP_per_service": "mean",
    }
)
df_mbs_sa3_gender_state

# Calculate additional variable to understand state level mbs datapoints
df_mbs_sa3_gender_state["%_People_had_service"] = (
    df_mbs_sa3_gender_state["No_of_patients"] / df_mbs_sa3_gender_state["ERP"] * 100
)
df_mbs_sa3_gender_state["Out_of_Pocket"] = (
    df_mbs_sa3_gender_state["Total_provider_fees_$"]
    - df_mbs_sa3_gender_state["Total_mbs_paid_$"]
)
df_mbs_sa3_gender_state["No_of_service_per_person"] = (
    df_mbs_sa3_gender_state["No_of_services"]
    / df_mbs_sa3_gender_state["No_of_patients"]
)
df_mbs_sa3_gender_state["OOP_per_person_annual"] = (
    df_mbs_sa3_gender_state["No_of_service_per_person"]
    * df_mbs_sa3_gender_state["Patient_OOP_per_service"]
)
df_mbs_sa3_gender_state["Total_Provider_pp_per_service"] = df_mbs_sa3_gender_state[
    "Total_provider_fees_$"
] / (
    df_mbs_sa3_gender_state["No_of_patients"]
    * df_mbs_sa3_gender_state["No_of_service_per_person"]
)
df_mbs_sa3_gender_state["MBS_Rebate_pp_per_service"] = df_mbs_sa3_gender_state[
    "Total_mbs_paid_$"
] / (
    df_mbs_sa3_gender_state["No_of_patients"]
    * df_mbs_sa3_gender_state["No_of_service_per_person"]
)

df_mbs_sa3_gender_state.head(10)

# exporting MBS-Census-State-Year-Service Level 1 (No demographic Data)
df_mbs_sa3_gender_state.to_csv(
    os.path.join(
        path, "clean_datasets/cleaned_csv/mbs_census_state_year_service_1_gender.csv"
    )
)

## State-Year MBS Data by Age Groups

df_mbs_sa3_service_1_np_new["DemographicGroup"].unique()

df_mbs_sa3_age_group = df_mbs_sa3_service_1_np_new[
    df_mbs_sa3_service_1_np_new["DemographicGroup"].isin(
        ["0-24", "25-44", "45-64", "65+"]
    )
]
df_mbs_sa3_age_group.shape

df_mbs_agegroup_state = df_mbs_sa3_age_group.groupby(
    by=["Year", "StateTerritory", "Service", "DemographicGroup"], as_index=False
).agg(
    {
        "No_of_patients": "sum",
        "No_of_services": "sum",
        "Total_mbs_paid_$": "sum",
        "Total_provider_fees_$": "sum",
        "ERP": "sum",
        "Patient_OOP_per_service": "mean",
    }
)
df_mbs_agegroup_state

# Calculate additional variable to understand state level mbs datapoints
df_mbs_agegroup_state["%_People_had_service"] = (
    df_mbs_agegroup_state["No_of_patients"] / df_mbs_agegroup_state["ERP"] * 100
)
df_mbs_agegroup_state["Out_of_Pocket"] = (
    df_mbs_agegroup_state["Total_provider_fees_$"]
    - df_mbs_agegroup_state["Total_mbs_paid_$"]
)

df_mbs_agegroup_state["No_of_service_per_person"] = (
    df_mbs_agegroup_state["No_of_services"] / df_mbs_agegroup_state["No_of_patients"]
)
df_mbs_sa3_gender_state["OOP_per_person_annual"] = (
    df_mbs_sa3_gender_state["No_of_service_per_person"]
    * df_mbs_sa3_gender_state["Patient_OOP_per_service"]
)
df_mbs_agegroup_state["Total_Provider_pp_per_service"] = df_mbs_agegroup_state[
    "Total_provider_fees_$"
] / (
    df_mbs_agegroup_state["No_of_patients"]
    * df_mbs_agegroup_state["No_of_service_per_person"]
)
df_mbs_agegroup_state["MBS_Rebate_pp_per_service"] = df_mbs_agegroup_state[
    "Total_mbs_paid_$"
] / (
    df_mbs_agegroup_state["No_of_patients"]
    * df_mbs_agegroup_state["No_of_service_per_person"]
)

df_mbs_agegroup_state.head(10)

# exporting MBS-Census-State-Year-Service Level 1 (No demographic Data)
df_mbs_agegroup_state.to_csv(
    os.path.join(
        path, "clean_datasets/cleaned_csv/mbs_census_state_year_service_1_agegroup.csv"
    )
)
